+++ cognite/experimental/data_classes/__init__.py
@@ -4,4 +4,6 @@ from cognite.experimental.data_classes.contextualization import *
 from cognite.experimental.data_classes.extractionpipelineruns import *
 from cognite.experimental.data_classes.extractionpipelines import *
 from cognite.experimental.data_classes.functions import *
+from cognite.experimental.data_classes.transformation_schedules import *
+from cognite.experimental.data_classes.transformations import *
 from cognite.experimental.data_classes.types import *