+++ cognite/experimental/data_classes/transformations.py
@@ -1,0 +1,261 @@
+from cognite.client.data_classes._base import *
+
+
+class TransformationDestination:
+    """TransformationDestination has static methods to define the target resource type of a transformation"""
+
+    def __init__(self, type: str = None):
+        self.type = type
+
+    @staticmethod
+    def assets():
+        """To be used when the transformation is meant to produce assets."""
+        return TransformationDestination("assets")
+
+    @staticmethod
+    def timeseries():
+        """To be used when the transformation is meant to produce time series."""
+        return TransformationDestination("timeseries")
+
+    @staticmethod
+    def assethierarchy():
+        """To be used when the transformation is meant to produce asset hierarchies."""
+        return TransformationDestination("assethierarchy")
+
+    @staticmethod
+    def events():
+        """To be used when the transformation is meant to produce events."""
+        return TransformationDestination("events")
+
+    @staticmethod
+    def datapoints():
+        """To be used when the transformation is meant to produce numeric data points."""
+        return TransformationDestination("datapoints")
+
+    @staticmethod
+    def stringdatapoints():
+        """To be used when the transformation is meant to produce string data points."""
+        return TransformationDestination("stringdatapoints")
+
+    @staticmethod
+    def sequences():
+        """To be used when the transformation is meant to produce sequences."""
+        return TransformationDestination("sequences")
+
+    @staticmethod
+    def files():
+        """To be used when the transformation is meant to produce files."""
+        return TransformationDestination("files")
+
+    @staticmethod
+    def labels():
+        """To be used when the transformation is meant to produce labels."""
+        return TransformationDestination("labels")
+
+    @staticmethod
+    def relationships():
+        """To be used when the transformation is meant to produce relationships."""
+        return TransformationDestination("relationships")
+
+    @staticmethod
+    def raw(database: str = "", table: str = ""):
+        """To be used when the transformation is meant to produce raw table rows.
+
+        Args:
+            database (str): database name of the target raw table.
+            table (str): name of the target raw table
+
+        Returns:
+            TransformationDestination pointing to the target table
+        """
+        return RawTable(type="raw_table", raw_type="plain_raw", database=database, table=table)
+
+
+class RawTable(TransformationDestination):
+    def __init__(self, type: str = None, raw_type: str = None, database: str = None, table: str = None):
+        super().__init__(type=type)
+        self.rawType = raw_type
+        self.database = database
+        self.table = table
+
+
+class OidcCredentials:
+    def __init__(
+        self,
+        client_id: str = None,
+        client_secret: str = None,
+        scopes: str = None,
+        token_uri: str = None,
+        cdf_project_name: str = None,
+    ):
+
+        self.client_id = client_id
+        self.client_secret = client_secret
+        self.scopes = scopes
+        self.token_uri = token_uri
+        self.cdf_project_name = cdf_project_name
+
+
+class TransformationJobBlockade:
+    def __init__(self, reason: str = None, created_time: Optional[int] = None):
+        self.reason = reason
+        self.created_time = created_time
+
+
+class Transformation(CogniteResource):
+    """The transformations resource allows transforming data in CDF.
+
+    Args:
+        id (int): A server-generated ID for the object.
+        external_id (str): The external ID provided by the client. Must be unique for the resource type.
+        name (str): The name of the Transformation.
+        query (str): SQL query of the transformation.
+        destination (TransformationDestination): see TransformationDestination for options.
+        conflict_mode (str): What to do in case of id collisions: either "abort", "upsert", "update" or "delete"
+        is_public (bool): Indicates if the transformation is visible to all in project or only to the owner.
+        ignore_null_fields (bool): Indicates how null values are handled on updates: ignore or set null.
+        source_api_key (str): Configures the transformation to authenticate with the given api key on the source.
+        destination_api_key (str): Configures the transformation to authenticate with the given api key on the destination.
+        source_oidc_credentials (Optional[OidcCredentials]): Configures the transformation to authenticate with the given oidc credentials key on the destination.
+        destination_oidc_credentials (Optional[OidcCredentials]): Configures the transformation to authenticate with the given oidc credentials on the destination.
+        created_time (int): The number of milliseconds since 00:00:00 Thursday, 1 January 1970, Coordinated Universal Time (UTC), minus leap seconds.
+        last_updated_time (int): The number of milliseconds since 00:00:00 Thursday, 1 January 1970, Coordinated Universal Time (UTC), minus leap seconds.
+        owner (str): Owner of the transformation: requester's identity.
+        owner_is_current_user (bool): Indicates if the transformation belongs to the current user.
+        has_source_api_key (bool): Indicates if the transformation is configured with a source api key.
+        has_destination_api_key (bool): Indicates if the transformation is configured with a destination api key.
+        has_source_oidc_credentials (bool): Indicates if the transformation is configured with a source oidc credentials set.
+        has_destination_oidc_credentials (bool): Indicates if the transformation is configured with a destination oidc credentials set.
+        cognite_client (CogniteClient): The client to associate with this object.
+    """
+
+    def __init__(
+        self,
+        id: int = None,
+        external_id: str = None,
+        name: str = None,
+        query: str = None,
+        destination: TransformationDestination = None,
+        conflict_mode: str = None,
+        is_public: bool = True,
+        ignore_null_fields: bool = False,
+        source_api_key: str = None,
+        destination_api_key: str = None,
+        source_oidc_credentials: Optional[OidcCredentials] = None,
+        destination_oidc_credentials: Optional[OidcCredentials] = None,
+        created_time: Optional[int] = None,
+        last_updated_time: Optional[int] = None,
+        owner: str = None,
+        owner_is_current_user: bool = True,
+        has_source_api_key: Optional[bool] = None,
+        has_destination_api_key: Optional[bool] = None,
+        has_source_oidc_credentials: Optional[bool] = None,
+        has_destination_oidc_credentials: Optional[bool] = None,
+        cognite_client=None,
+    ):
+        self.id = id
+        self.external_id = external_id
+        self.name = name
+        self.query = query
+        self.destination = destination
+        self.conflict_mode = conflict_mode
+        self.is_public = is_public
+        self.ignore_null_fields = ignore_null_fields
+        self.source_api_key = source_api_key
+        self.has_source_api_key = has_source_api_key or source_api_key is not None
+        self.destination_api_key = destination_api_key
+        self.has_destination_api_key = has_destination_api_key or destination_api_key is not None
+        self.source_oidc_credentials = source_oidc_credentials
+        self.has_source_oidc_credentials = has_source_oidc_credentials or source_oidc_credentials is not None
+        self.destination_oidc_credentials = destination_oidc_credentials
+        self.has_destination_oidc_credentials = (
+            has_destination_oidc_credentials or destination_oidc_credentials is not None
+        )
+        self.created_time = created_time
+        self.last_updated_time = last_updated_time
+        self.owner = owner
+        self.owner_is_current_user = owner_is_current_user
+        self._cognite_client = cognite_client
+
+    @classmethod
+    def _load(cls, resource: Union[Dict, str], cognite_client=None):
+        instance = super(Transformation, cls)._load(resource, cognite_client)
+        if isinstance(instance.destination, Dict):
+            snake_dict = {utils._auxiliary.to_snake_case(key): value for (key, value) in instance.destination.items()}
+            if instance.destination.get("type") == "raw_table":
+                instance.destination = RawTable(**snake_dict)
+            else:
+                instance.destination = TransformationDestination(**snake_dict)
+        return instance
+
+    def __hash__(self):
+        return hash(self.external_id)
+
+
+class TransformationUpdate(CogniteUpdate):
+    """Changes applied to transformation
+
+    Args:
+        id (int): A server-generated ID for the object.
+        external_id (str): External Id provided by client. Should be unique within the project.
+    """
+
+    class _PrimitiveTransformationUpdate(CognitePrimitiveUpdate):
+        def set(self, value: Any) -> "TransformationUpdate":
+            return self._set(value)
+
+    @property
+    def name(self):
+        return TransformationUpdate._PrimitiveTransformationUpdate(self, "name")
+
+    @property
+    def destination(self):
+        return TransformationUpdate._PrimitiveTransformationUpdate(self, "destination")
+
+    @property
+    def conflict_mode(self):
+        return TransformationUpdate._PrimitiveTransformationUpdate(self, "conflictMode")
+
+    @property
+    def query(self):
+        return TransformationUpdate._PrimitiveTransformationUpdate(self, "query")
+
+    @property
+    def source_oidc_credentials(self):
+        return TransformationUpdate._PrimitiveTransformationUpdate(self, "sourceOidcCredentials")
+
+    @property
+    def destination_oidc_credentials(self):
+        return TransformationUpdate._PrimitiveTransformationUpdate(self, "destinationOidcCredentials")
+
+    @property
+    def source_api_key(self):
+        return TransformationUpdate._PrimitiveTransformationUpdate(self, "sourceApiKey")
+
+    @property
+    def destination_api_key(self):
+        return TransformationUpdate._PrimitiveTransformationUpdate(self, "destinationApiKey")
+
+    @property
+    def is_public(self):
+        return TransformationUpdate._PrimitiveTransformationUpdate(self, "isPublic")
+
+    @property
+    def ignore_null_fields(self):
+        return TransformationUpdate._PrimitiveTransformationUpdate(self, "ignoreNullFields")
+
+
+class TransformationList(CogniteResourceList):
+    _RESOURCE = Transformation
+    _UPDATE = TransformationUpdate
+
+
+class TransformationFilter(CogniteFilter):
+    """No description.
+
+    Args:
+        include_public (bool): Whether public transformations should be included in the results. The default is true.
+    """
+
+    def __init__(self, include_public: bool = True):
+        self.include_public = include_public